packages/playfile-core/pyproject.toml
[project]
name = "playfile-core"
version = "0.1.0"
description = "Core configuration system for Playfile"
readme = "README.md"
authors = [
    { name = "Sofiane Djerbi", email = "sofiane.djerbi38@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dacite>=1.9.2",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
